+++ Changelog.md
@@ -4,15 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-- button for navigation to starting screen
-    
+  
 
-## [unreleased]
+## [v0.1.0] - 2024-07-16
 
 ### Added
 
 - results screen
 - widgets to display results properly
+- button for navigation to starting screen
 
 ## [v0.0.3] - 2024-07-02